+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 
 - Crash on Android (Termux) due to unknown user runtime directory
+- Crash due to misconfigured or unavailable audio backend
 
 ## [1.0.0] - 2023-12-16